+++ xlive/Blam/Engine/Networking/logic/life_cycle_manager.h
