+++ xlive/Blam/Engine/tag_files/files_windows.h
